+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- The error returned by `Task.JoinWithoutPanicking()` now wraps the panic value if the panic value is an error
+
 ## [1.1.1] - 2023-11-29
 
 ### Changed